libidset: relocate veb source
Problem: the veb class's only user is libidset, and idset is
the preferred interface since it is safer.

Move the veb class to libidset.
